requirements.txt
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.4.4
CacheControl==0.12.6
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.0.8
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-paypal==0.2.5
django-reflinks==0.5.1
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.0.1
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==6.2.2
progress==1.5
pycodestyle==2.5.0
pyparsing==2.4.6
python-decouple==3.1
python3-openid==3.1.0
pytoml==0.1.21
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
six==1.14.0
South==1.0.2
sqlparse==0.2.4
stripe==2.50.0
urllib3==1.24.2
webencodings==0.5.1
appdirs==1.4.3
asgiref==3.2.10
autopep8==1.4.4
CacheControl==0.12.6
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==2.2
django-allauth==0.39.1
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-paypal==0.2.5
django-reflinks==0.5.1
django-user-visit==0.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
oauthlib==3.0.1
packaging==20.3
pep517==0.8.2
pep8==1.7.1
Pillow==6.2.2
progress==1.5
pycodestyle==2.5.0
pyparsing==2.4.6
python-decouple==3.1
python3-openid==3.1.0
pytoml==0.1.21
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
retrying==1.3.3
six==1.14.0
South==1.0.2
sqlparse==0.2.4
stripe==2.50.0
ua-parser==0.10.0
urllib3==1.24.2
user-agents==2.2.0
webencodings==0.5.1
